feat: implement useRandomGlow hook with shadow color utility and refactor color functions
